[DEV-667] restructured remaining mocks directories
