+++ src/payment/models/buy-crypto/services/__tests__/buy-crypto.service.spec.ts
@@ -7,7 +7,7 @@ import { CryptoInputType } from 'src/payment/models/crypto-input/crypto-input.en
 import { createCustomCryptoInput } from 'src/payment/models/crypto-input/__mocks__/crypto-input.entity.mock';
 import { CryptoRouteRepository } from 'src/payment/models/crypto-route/crypto-route.repository';
 import { CryptoRouteService } from 'src/payment/models/crypto-route/crypto-route.service';
-import { createCustomCryptoRouteHistory } from 'src/payment/models/crypto-route/dto/__tests__/mock/crypto-route-history.dto.mock';
+import { createCustomCryptoRouteHistory } from 'src/payment/models/crypto-route/dto/__mocks__/crypto-route-history.dto.mock';
 import { SettingService } from 'src/shared/models/setting/setting.service';
 import { TestSharedModule } from 'src/shared/test.shared.module';
 import { UserService } from 'src/user/models/user/user.service';
