+++ src/payment/models/sell/__mocks__/sell.entity.mock.ts
@@ -1,6 +1,6 @@
 import { createDefaultBankAccount } from 'src/payment/models/bank-account/__mocks__/bank-account.entity.mock';
 import { createDefaultDeposit } from 'src/payment/models/deposit/__mocks__/deposit.entity.mock';
-import { createDefaultFiat } from 'src/shared/models/fiat/__tests__/mock/fiat.entity.mock';
+import { createDefaultFiat } from 'src/shared/models/fiat/__mocks__/fiat.entity.mock';
 import { createDefaultUser } from 'src/user/models/user/__mocks__/user.entity.mock';
 import { Sell } from '../sell.entity';
